feature: started builtin new lexer
I started builtin a new lexer in case we cannot solve the problem
in the existing one preventing correct function execution. This
is unlikely but in case, as I was stuck on the current lexer
I started builtin a alternative from the ground up now that
we understand better how the Bash lexer works.