feature: added parsing of redirection after cmd_parsing
We where previously not parsing the redirection if it came
after a command between () because we where assuming
that redirection was always happening at the simple_cmd
level. However this assumption was wrong. This is no fixed
and we can handle command like:
(cat | grep in) <infile
However we need to be aware that probably some safeguard are
missing and the current implementation might be a little bit
redundant in the code, it could probably be improved.